// client/src/typescript-fetch-client/models/ModelsComment.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Server API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { GormDeletedAt } from './GormDeletedAt';
import {
    GormDeletedAtFromJSON,
    GormDeletedAtFromJSONTyped,
    GormDeletedAtToJSON,
    GormDeletedAtToJSONTyped,
} from './GormDeletedAt';
import type { ModelsUser } from './ModelsUser';
import {
    ModelsUserFromJSON,
    ModelsUserFromJSONTyped,
    ModelsUserToJSON,
    ModelsUserToJSONTyped,
} from './ModelsUser';

/**
 * 
 * @export
 * @interface ModelsComment
 */
export interface ModelsComment {
    /**
     * 
     * @type {string}
     * @memberof ModelsComment
     */
    createdAt?: string;
    /**
     * 
     * @type {GormDeletedAt}
     * @memberof ModelsComment
     */
    deletedAt?: GormDeletedAt;
    /**
     * 
     * @type {number}
     * @memberof ModelsComment
     */
    id?: number;
    /**
     * 
     * @type {number}
     * @memberof ModelsComment
     */
    taskId?: number;
    /**
     * 
     * @type {string}
     * @memberof ModelsComment
     */
    text?: string;
    /**
     * 
     * @type {string}
     * @memberof ModelsComment
     */
    updatedAt?: string;
    /**
     * 
     * @type {ModelsUser}
     * @memberof ModelsComment
     */
    user?: ModelsUser;
    /**
     * 
     * @type {number}
     * @memberof ModelsComment
     */
    userId?: number;
}

/**
 * Check if a given object implements the ModelsComment interface.
 */
export function instanceOfModelsComment(value: object): value is ModelsComment {
    return true;
}

export function ModelsCommentFromJSON(json: any): ModelsComment {
    return ModelsCommentFromJSONTyped(json, false);
}

export function ModelsCommentFromJSONTyped(json: any, ignoreDiscriminator: boolean): ModelsComment {
    if (json == null) {
        return json;
    }
    return {
        
        'createdAt': json['created_at'] == null ? undefined : json['created_at'],
        'deletedAt': json['deleted_at'] == null ? undefined : GormDeletedAtFromJSON(json['deleted_at']),
        'id': json['id'] == null ? undefined : json['id'],
        'taskId': json['task_id'] == null ? undefined : json['task_id'],
        'text': json['text'] == null ? undefined : json['text'],
        'updatedAt': json['updated_at'] == null ? undefined : json['updated_at'],
        'user': json['user'] == null ? undefined : ModelsUserFromJSON(json['user']),
        'userId': json['user_id'] == null ? undefined : json['user_id'],
    };
}

export function ModelsCommentToJSON(json: any): ModelsComment {
    return ModelsCommentToJSONTyped(json, false);
}

export function ModelsCommentToJSONTyped(value?: ModelsComment | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'created_at': value['createdAt'],
        'deleted_at': GormDeletedAtToJSON(value['deletedAt']),
        'id': value['id'],
        'task_id': value['taskId'],
        'text': value['text'],
        'updated_at': value['updatedAt'],
        'user': ModelsUserToJSON(value['user']),
        'user_id': value['userId'],
    };
}